KIN-744 moved import from its own view to a modal
